refactor: restructure ai package
Here refactored ai package to better split server and client parts
and not leak babel to client.

This save ~100kB in our bundle size.
