+++ packages/ai/src/chains/command-detect/chain.server.ts
@@ -1,28 +1,11 @@
-import { z } from "zod";
 import type { Model as BaseModel, ModelMessage, Chain } from "../../types";
 import { formatPrompt } from "../../utils/format-prompt";
 import { prompt as promptSystemTemplate } from "./__generated__/command-detect.system.prompt";
 import { prompt as promptUserTemplate } from "./__generated__/command-detect.user.prompt";
 import { createErrorResponse } from "../../utils/create-error-response";
+import { type Context, type Response, ResponseSchema, name } from "./schema";
 
-/**
- * Command Detect Chain
- *
- * Given a prompt and a list of possible commands and descriptions, it returns an array of operations matching the prompt request.
- */
-
-export const name = "command-detect";
-
-export const ContextSchema = z.object({
-  // The prompt provides the original user request.
-  prompt: z.string(),
-  // Command name - description pairs.
-  commands: z.record(z.string(), z.string()),
-});
-export type Context = z.infer<typeof ContextSchema>;
-
-export const ResponseSchema = z.array(z.string());
-export type Response = z.infer<typeof ResponseSchema>;
+export { name };
 
 export const createChain = <ModelMessageFormat>(): Chain<
   BaseModel<ModelMessageFormat>,
