--- packages/ai/src/chains/command-detect/index.server.ts
@@ -1,1 +1,0 @@
-export * from "./chain";
