--- packages/ai/src/chains/command-detect/index.ts
@@ -1,1 +1,0 @@
-export { name, type Response, ResponseSchema } from "./chain";
