--- packages/ai/src/chains/copywriter/index.server.ts
@@ -1,1 +1,0 @@
-export * from "./chain";
