--- packages/ai/src/chains/copywriter/index.ts
@@ -1,8 +1,0 @@
-export {
-  name,
-  type Response,
-  ResponseSchema,
-  type TextInstance,
-  TextInstanceSchema,
-  collectTextInstances,
-} from "./chain";
