+++ packages/ai/src/chains/operations/delete-instance.server.ts
@@ -1,0 +1,9 @@
+import type { aiOperation, wsOperation } from "./delete-instance";
+
+export { name } from "./delete-instance";
+
+export type { wsOperation };
+
+export const aiOperationToWs = async (
+  operation: aiOperation
+): Promise<wsOperation> => operation;
