+++ packages/ai/src/chains/operations/generate-insert-template.ts
@@ -1,6 +1,5 @@
 import { z } from "zod";
 import { WsEmbedTemplate, idAttribute } from "@webstudio-is/react-sdk";
-import { jsxToTemplate } from "../../utils/jsx-to-template.server";
 
 // Currently this operation is used is a separate LLM call after the main one has returned an insert-instance operation. This is to produce better results.
 // Effectively the aiOperation from this module is not used. The separate chain will call the LLM and we use the resulting completion to construct an aiOperation by hand
@@ -35,14 +34,3 @@ export const wsOperation = z.object({
   template: WsEmbedTemplate,
 });
 export type wsOperation = z.infer<typeof wsOperation>;
-
-export const aiOperationToWs = async (
-  operation: aiOperation
-): Promise<wsOperation> => {
-  return {
-    operation: "insertTemplate",
-    addTo: operation.addTo,
-    addAtIndex: operation.addAtIndex,
-    template: await jsxToTemplate(operation.jsx),
-  };
-};
