--- packages/ai/src/chains/template-generator/index.server.ts
@@ -1,1 +1,0 @@
-export * from "./chain";
