--- packages/ai/src/chains/template-generator/index.ts
@@ -1,1 +1,0 @@
-export { name, type Response, ResponseSchema } from "./chain";
