+++ packages/ai/src/image.ts
@@ -1,4 +1,3 @@
-import { traverseTemplate } from "@webstudio-is/jsx-utils";
 import type {
   EmbedTemplateInstance,
   EmbedTemplateProp,
@@ -130,6 +129,18 @@ const queryImageAndMutateInstance = async (
       property: "objectFit",
       value: { type: "keyword", value: "cover" },
     });
+  }
+};
+
+const traverseTemplate = (
+  template: WsEmbedTemplate,
+  fn: (node: WsEmbedTemplate[number]) => void
+) => {
+  for (const node of template) {
+    fn(node);
+    if (node.type === "instance") {
+      traverseTemplate(node.children, fn);
+    }
   }
 };
