+++ packages/ai/src/index.ts
@@ -1,8 +1,8 @@
 /** Chains */
-export * as commandDetect from "./chains/command-detect/index";
-export * as copywriter from "./chains/copywriter/index";
-export * as operations from "./chains/operations/index";
-export * as templateGenerator from "./chains/template-generator/index";
+export * as commandDetect from "./chains/command-detect/schema";
+export * as copywriter from "./chains/copywriter/schema";
+export * as operations from "./chains/operations/schema";
+export * as templateGenerator from "./chains/template-generator/schema";
 
 /** Utils */
 export * from "./utils/handle-ai-request";
